Removing activity 'inicio'/migration to main
