+++ app/src/main/java/com/bandtec/finfamily/MainActivity.kt
@@ -1,12 +1,25 @@
 package com.bandtec.finfamily
 
+import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
+
+        logininicio.setOnClickListener {
+            val intent = Intent(this, Login::class.java)
+            // start your next activity
+            startActivity(intent)
+        }
+        cadastroinicio.setOnClickListener {
+            val intent = Intent(this, Extract::class.java)
+            // start your next activity
+            startActivity(intent)
+        }
     }
 }
