--- app/src/main/res/layout/activity_inicio.xml
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/linearLayout"
-    android:layout_width="fill_parent"
-    android:layout_height="fill_parent"
-    android:orientation="vertical"
-    android:gravity="bottom"
-    android:background="@drawable/fundo"
-
-
-    tools:context=".Inicio">
-
-    <!-- fundo -->
-    <!--    <View-->
-    <!--        android:id="@+id/fundo"-->
-    <!--        android:layout_width="416dp"-->
-    <!--        android:layout_height="729dp"-->
-    <!--        android:layout_alignParentLeft="true"-->
-    <!--        android:layout_alignParentTop="true"-->
-    <!--        android:layout_marginTop="1dp"-->
-    <!--        android:layout_marginBottom="1dp"-->
-    <!--        android:background="@drawable/fundoinicial"-->
-    <!--        android:elevation="4dp"-->
-    <!--        app:layout_constraintBottom_toBottomOf="parent"-->
-    <!--        app:layout_constraintEnd_toEndOf="parent"-->
-    <!--        app:layout_constraintStart_toStartOf="parent"-->
-    <!--        app:layout_constraintTop_toTopOf="parent" />-->
-
-    <!-- FinFamily -->
-
-    <TextView
-        android:id="@+id/finfamily"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="center"
-        android:layout_marginBottom="225dp"
-        android:text="@string/finfamily"
-        android:textAppearance="@style/finfamily" />
-
-
-    <LinearLayout
-        android:id="@+id/linearLayout1"
-        android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:layout_marginBottom="50dp"
-        android:layout_marginEnd="20dp"
-        android:layout_marginStart="20dp"
-        >
-
-        <Button
-            android:id="@+id/cadastroinicio"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_alignParentRight="true"
-            android:layout_marginBottom="10dp"
-            android:background="@drawable/cadastroinicial"
-            android:text="Create Account"/>
-
-        <Button
-            android:id="@+id/logininicio"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:background="@drawable/buttonlogininicio"
-            android:text="Login"/>
-
-    </LinearLayout>
-
-
-</LinearLayout>
